requirements.txt
# Flask and extensions
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.0
email-validator>=2.0.0

# OAuth authentication
Authlib>=1.3.0
requests>=2.31.0

# Password hashing
bcrypt>=4.1.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
hypothesis>=6.100.0

# Environment variables
python-dotenv>=1.0.0

# Production server
gunicorn>=21.0.0
